change default expiry tome to 1 day
So that the invoice still has a change to be paid if the creator of the invoice gets the payer to open a channel with him. The payer will want to open a channel that is higher than the amount of the invoice. It is possible that if a new invoice is generate, a new amount will need to be generated, and asking to generate a new invoice makes it difficult for automated LN sending services like bullbitcoin.com (me) to pay out in LN. If there is no route to the user's node, I will open a channel and pay once there are 6 confirmations, but if the bolt11 invoice expiry is 1h, it will happen very often that it will take more than 60 minutes for 6 confirmations, and then the user will have to go back to the application and put a new bolt11 for his withdrawal, which is highly inconvenient.

Users paying the same bolt11 twice is not an issue, this is already possible with Bitcoin where users frequently pay invoices twice. Anybody using Bitcoin knows, or should know, how to deal with this at the applicaiton layer.